move to namespace dt::df::core. replaced prototypes with real ones
