+++ core/include/dt/df/core/types.hpp
@@ -1,19 +1,20 @@
 #pragma once
 #include <memory>
 #include <string>
+#include <unordered_map>
 #include <vector>
 #include <nlohmann/json.hpp>
-namespace dt::df
+namespace dt::df::core
 {
 class IGraphManager;
 class BaseNode;
-class BaseSlot;
+class CoreSlot;
 using NodeId = int;
 using NodeKey = std::string;
 using NodePtr = std::shared_ptr<BaseNode>;
-using SlotPtr = std::shared_ptr<BaseSlot>;
-using Slots = std::vector<SlotPtr>;
+using SlotPtr = std::shared_ptr<CoreSlot>;
 using SlotId = int;
+using SlotMap = std::unordered_map<SlotId /*local_id*/, SlotPtr>;
 using SlotName = std::string;
 using SlotKey = std::string_view;
 enum class SlotType
