--- core/src/prototype_slot.cpp
@@ -1,4 +1,0 @@
-#include "dt/df/core/prototype_slot.hpp"
-
-namespace dt::df::core
-{} // namespace dt::df::core
